Fix Windows Compatibility Issues
- Resolve `file in use` error in `TestDB_Init_FileError`.
- Move permission error tests to a Unix-specific test file.
- Isolate `renameFile` in a Unix-specific file to prevent redefinition on Windows.
